Add functionality for drawing images
This allows images to be loaded and drawn to the canvas by means of the
whole image, a rotated version of the image, a scaled version of the
image, and a partial version of the image. Image dimensions can also
be procured.

Further, this fixes an issue where setTimeout() was not working as
expected in resourcesLoaded().