Listen for mouse events on the canvas element
This makes it such that mouse events are only fired inside of the
canvas element itself instead of the entire document. This ensures
that the mouse offset will be not result in strange behavior when the
mouse leaves the canvas. When the mouse leaves the canvas, the events
stop firing.