Check for continuous holding of mouse buttons
This implements isMouseButtonUp() and isMouseButtonDown(). The former
is used to continuously check if a particular mouse button is being
held down, and the latter to continuously check if a button is not
being held down (to check if it is up).

For one-time button presses, the methods isMouseButtonPressed() and
isMouseButtonReleased() should be used instead.