Check for any key or mouse button
The string of "any" can be used in place of a key code or button name
to check whether or not the corresponding event has any matches. This
enables games to display "press any key to continue". For cases where
this is used against checking if a key or button is up, the result is
always true, as there is always at least one of each that is up.