Change which mouse events should be canceled
Only the scroll wheel and and right-clicking events should be canceled.
The scroll wheel should be canceled so the rest of the page does not
scroll when it is used, and right-clicking should be canceled to
prevent the browser's context menu from appearing.

Moving the mouse, clicking the left and middle mouse buttons, and
releasing the mouse buttons should not be canceled, so as to allow
focus to be sent in the event that games are played within iframes.

This fixes #7.