Load existing font faces
Add an optional Boolean to Momo.loadFont() that tells Momo to use a
font face that has been defined in CSS or is already known to the
browser (such as Web safe fonts). If set to true, the font_name
variable will be used as the name of the font face.

Further, remove the "font_" prefix from font names, and remove the file
and element properties from the font's object literal.

This closes #9.