Check for errors when loading resources
Changed Momo.loadFont(), Momo.loadBitmap(), and Momo.loadSample() to
return promises so error checking can be performed. These methods
should be called within async functions using the await keyword.

Momo.loadFont() returns an object literal containing font data if the
file passed to it can be loaded. It otherwise returns false.

Momo.loadBitmap() returns an object literal containing bitmap data if
the image file passed to it loads successfully. If the image fails to
load, it returns false.

Momo.loadSample() returns an object literal containing sample data if
the file passed to it can be loaded. False is returned if the browser
does not support the file format, or if the file fails to load.

This closes #12.