Fix drawBitmap() and fix transformations
The method drawBitmap() has been updated to actually draw the bitmap in
the proper pixel position, and in the proper scaled resolution. Also,
the shader now works in pixel space and converts it to clip space.

Also, transformations work again thanks to newly-added matrix methods.
One must call createTransformation() to get an identity matrix, and
pass its return to the transformation methods. Finally, the method
applyTransformation() must be called to apply everything.