Restructure createGameLoop()
Treat updating and rendering as two distinct phases. Instead of using a
monolithic anonymous function for game loops, pass a procedure for
updating and rendering separately.

Use requestAnimationFrame() to let the browser decide when to best
repaint the canvas for the render procedure. The frame-rate is no
longer controllable here, but will likely target 60 Hz.

The interval for updating remains controllable. This commit should
allow for smoother and more consistent repaints.