Rewrite createTintedBitmap() to work with WebGL
The way this method used to work was by drawing a tinted version of the
bitmap and saving it to a sub-canvas. Sub-canvases worked well in the
context of Canvas 2D, but are irrelevant in WebGL.

Now, this method works by initializing bitmaps with a dummy tint of -1
for all color channels. Rather than modify the source texture, this
method creates a new bitmap object with the given tint as a part of the
object literal. Then, when drawing bitmaps, a check is performed to see
if the given bitmap has a tint applied to it. If it does, it combines
that tint with the tint passed to drawConsolidatedBitmap(), thus mixing
the two. If only the dummy tint exists, the default tint is applied.

Tints are always passed to the shader, no matter the drawing method, so
this ends up leveraging an existing system, and comes at no cost to
performance.