Limit game loop to 60 Hz
Depending on the implementation, requestAnimationFrame() may match the
refresh rate of the user's display. In most cases, this would equal 60
Hz, but it could be greater in higher refresh rate displays.

Playing a game on a higher refresh rate display may result in the game
loop being executed more frequently than anticipated, resulting in
animations playing too quickly, for example.

To address the above, this commit limits the game loop to 60 Hz.